refactor!: rename the package and every in-repo reference to comfy-mcp
The name shipping publicly is `comfy-mcp`, not `comfy-local-mcp`: users expect
a Comfy MCP to drive any ComfyUI instance, and the cloud MCP is the remote-HTTP
version of the same idea, so "local" does not belong in the name. Doing it now,
before the first PyPI publish, is what keeps it free — nothing has been
released under the old name, so there is no distribution to alias, no
deprecation shim to carry, and no user data to migrate.

Renamed surfaces:

- `[project] name` -> `comfy-mcp`, and the `description` drops its leading
  "Local".
- `[project.scripts]` -> `comfy-mcp = "comfy_mcp.server:main"`. This is the
  user-visible one: it is the `"command"` in every MCP client config, so all
  five README client snippets (Claude Code CLI + `.mcp.json`, Claude Desktop,
  Cursor, non-default address) move with it.
- `src/comfy_local_mcp/` -> `src/comfy_mcp/`, plus the imports in the 25 test
  modules and the `MCPServer(...)` name the server advertises over the wire.
- The opt-in failure log moves with the package: its directory leaf on all
  three platforms (`~/Library/Application Support/comfy-mcp/failures.jsonl`
  and the Windows/Linux twins), its logger name, and its env toggle
  `COMFY_MCP_DEBUG_LOG`.
- Docs, issue templates, workflow comments and repository URLs
  (`Comfy-Org/comfy-mcp`).

`COMFY_LOCAL_URL` and the "local" wording in prose and tool descriptions are
deliberately untouched; they are a separate piece of work.
