+++ .github/workflows/secret-scanning.yml
@@ -16,7 +16,7 @@ name: Secret Scanning
 # emails the last person to touch this file. A dormant repo — the case the
 # rescan exists for — is therefore exactly the case where it stops silently.
 # Run it by hand after a long quiet spell:
-#   gh workflow run secret-scanning.yml -R Comfy-Org/comfy-local-mcp
+#   gh workflow run secret-scanning.yml -R Comfy-Org/comfy-mcp
 
 on:
   pull_request:
