+++ NOTICE
@@ -1,4 +1,4 @@
-comfy-local-mcp
+comfy-mcp
 Copyright (c) Comfy Org
 
 This product includes software developed at Comfy Org
