--- src/comfy_local_mcp/__init__.py
@@ -1,3 +1,0 @@
-"""comfy-local-mcp: a thin MCP wrapper over comfy-cli."""
-
-__version__ = "0.4.0"
