+++ src/comfy_mcp/__init__.py
@@ -1,0 +1,3 @@
+"""comfy-mcp: a thin MCP wrapper over comfy-cli."""
+
+__version__ = "0.4.0"
