+++ tests/test_discovery.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 import pytest
 from conftest import envelope
 
-from comfy_local_mcp import server
+from comfy_mcp import server
 
 
 def test_search_nodes_argv(patched_run):
