+++ tests/test_generate.py
@@ -22,7 +22,7 @@ import asyncio
 import pytest
 from conftest import _OK_STREAM, _RecordingCtx
 
-from comfy_local_mcp import server
+from comfy_mcp import server
 
 
 def test_generate_image_streams_and_maps_command(patched_stream):
