+++ tests/test_network_exposure.py
@@ -41,7 +41,7 @@ from mcp.server.elicitation import (
     DeclinedElicitation,
 )
 
-from comfy_local_mcp import server
+from comfy_mcp import server
 
 
 def _launch(*args, **kwargs):
