+++ tests/test_partner_auth.py
@@ -18,7 +18,7 @@ import asyncio
 
 import pytest
 
-from comfy_local_mcp import server
+from comfy_mcp import server
 
 
 def _partner_error_envelope() -> dict:
