+++ tests/test_partner_generate.py
@@ -33,7 +33,7 @@ from mcp.server.elicitation import (
     DeclinedElicitation,
 )
 
-from comfy_local_mcp import server
+from comfy_mcp import server
 
 
 def _generate(*args, **kwargs):
