+++ tests/test_run_template.py
@@ -37,7 +37,7 @@ from mcp.server.elicitation import (
     DeclinedElicitation,
 )
 
-from comfy_local_mcp import server
+from comfy_mcp import server
 
 
 def _run_template(*args, **kwargs):
