+++ tests/test_switch_version.py
@@ -34,7 +34,7 @@ from mcp.server.elicitation import (
     DeclinedElicitation,
 )
 
-from comfy_local_mcp import server
+from comfy_mcp import server
 
 
 def _switch(*args, **kwargs):
