+++ tests/test_templates.py
@@ -23,7 +23,7 @@ import os
 import pytest
 from conftest import envelope
 
-from comfy_local_mcp import server
+from comfy_mcp import server
 
 # A representative slice of the real comfy-cli `templates ls` payload shape.
 ROWS = [
